Made edit user endpoint

+++ WebApi/Controllers/UsersController.cs
@@ -1,3 +1,4 @@
+using Auth;
 using Auth.Authorization;
 using Auth.Authorization.Attributes;
 using AutoMapper;
@@ -46,7 +47,7 @@ public class UsersController : Controller
     }
 
     [Authorize(Roles = ApiRoles.Webmaster)]
-    [HttpPost("Create")]
+    [HttpPost]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult> Create([FromBody] CreateUserRequestModel model)
@@ -64,17 +65,45 @@ public class UsersController : Controller
     }
 
     [Authorize(Roles = ApiRoles.Webmaster)]
-    [HttpDelete("Delete/{id}")]
+    [HttpDelete("{id}")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult> Delete([FromRoute] string id)
     {
         var user = await _userManager.FindByIdAsync(id);
-        if (user == null) return NotFound();
+        if (user == null || user.Id == AuthConstants.GhostUserId) return NotFound();
+
+        if (user.Id == AuthConstants.DefaultWebmasterId) return BadRequest("Cannot delete default webmaster");
 
         var deleteResult = await _userManager.DeleteAsync(user);
         if (!deleteResult.Succeeded) return BadRequest(deleteResult.Errors);
+
+        return NoContent();
+    }
+
+    [Authorize(Roles = ApiRoles.Webmaster)]
+    [HttpPatch("{id}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<ActionResult> Edit([FromRoute] string id, [FromBody] EditUserRequestModel model)
+    {
+        var user = await _userManager.FindByIdAsync(id);
+        if (user == null || user.Id == AuthConstants.GhostUserId) return NotFound();
+
+        _mapper.Map(model, user);
+
+        var updateResult = await _userManager.UpdateAsync(user);
+
+        if (!updateResult.Succeeded) return BadRequest(updateResult.Errors);
+
+        if (model.Password != null)
+        {
+            var token = await _userManager.GeneratePasswordResetTokenAsync(user);
+            var resetResult = await _userManager.ResetPasswordAsync(user, token, model.Password);
+            if (!resetResult.Succeeded) return BadRequest(resetResult.Errors);
+        }
 
         return NoContent();
     }
